Refactor GRPC stream into own file
