+++ rust/processor/src/lib.rs
@@ -11,11 +11,12 @@
 #[macro_use]
 extern crate diesel;
 
+pub use config::IndexerGrpcProcessorConfig;
+
 mod config;
+pub mod grpc_stream;
 pub mod models;
 pub mod processors;
 pub mod schema;
 pub mod utils;
 pub mod worker;
-
-pub use config::IndexerGrpcProcessorConfig;
